# backend/api/openapi.json
{
 "openapi": "3.0.3",
 "info": {
  "title": "Aether Foundation API",
  "version": "0.1.0",
  "description": "Aether cloud/on-prem API: authenticated MQTT telemetry, persisted sensor history and Widget Studio."
 },
 "servers": [
  {
   "url": "http://localhost:8080",
   "description": "Loopback development only; use HTTPS origin in production."
  }
 ],
 "paths": {
...
  "/api/v1/auth/register": {
   "post": {
    "summary": "Development registration; production/onprem disabled",
    "operationId": "post__api_v1_auth_register",
    "security": [],
    "responses": {
     "201": {
      "description": "Success",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Account"
        }
       }
      }
     },
     "default": {
      "description": "Rejected; 400 invalid input, 401 credentials, 403 authorization/origin or detail=password_change_required, 404 inaccessible resource, 409 conflict/quota, 429 rate limit, 500 internal.",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/Registration"
       }
      }
     }
    },
    "parameters": [
     {
      "name": "Origin",
      "in": "header",
      "required": true,
      "schema": {
       "type": "string"
      },
      "description": "Must equal configured APP_ORIGIN exactly."
     }
    ]
   }
  },
  "/api/v1/auth/login": {
   "post": {
    "summary": "Login; sets HttpOnly refresh cookie",
    "operationId": "post__api_v1_auth_login",
    "security": [],
    "responses": {
     "200": {
      "description": "Success",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AuthResult"
        }
       }
      }
     },
     "default": {
      "description": "Rejected; 400 invalid input, 401 credentials, 403 authorization/origin or detail=password_change_required, 404 inaccessible resource, 409 conflict/quota, 429 rate limit, 500 internal.",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/Login"
       }
      }
     }
    },
    "parameters": [
     {
      "name": "Origin",
      "in": "header",
      "required": true,
      "schema": {
       "type": "string"
      },
      "description": "Must equal configured APP_ORIGIN exactly."
     }
    ]
   }
  },
  "/api/v1/auth/refresh": {
   "post": {
    "summary": "Rotate refresh cookie; reuse revokes session family",
    "operationId": "post__api_v1_auth_refresh",
    "security": [
     {
      "refreshCookie": []
     },
     {
      "developmentRefreshCookie": []
     }
    ],
    "responses": {
     "200": {
      "description": "Success",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AuthResult"
        }
       }
      }
     },
     "default": {
      "description": "Rejected; 400 invalid input, 401 credentials, 403 authorization/origin or detail=password_change_required, 404 inaccessible resource, 409 conflict/quota, 429 rate limit, 500 internal.",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "parameters": [
     {
      "name": "Origin",
      "in": "header",
      "required": true,
      "schema": {
       "type": "string"
      },
      "description": "Must equal configured APP_ORIGIN exactly."
     }
    ]
   }
  },
  "/api/v1/auth/logout": {
   "post": {
    "summary": "Revoke current session including access tokens",
    "operationId": "post__api_v1_auth_logout",
    "security": [
     {
      "bearerAuth": []
     }
    ],
    "responses": {
     "204": {
      "description": "Success"
     },
     "default": {
      "description": "Rejected; 400 invalid input, 401 credentials, 403 authorization/origin or detail=password_change_required, 404 inaccessible resource, 409 conflict/quota, 429 rate limit, 500 internal.",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "parameters": [
     {
      "name": "Origin",
      "in": "header",
      "required": true,
      "schema": {
       "type": "string"
      },
      "description": "Must equal configured APP_ORIGIN exactly."
     }
    ]
   }
  },
...
  "/api/v1/catalog": {
   "get": {
    "summary": "Gateway models and device profiles this deployment accepts; verified=true only for hardware-captured golden tests",
    "operationId": "get__api_v1_catalog",
    "security": [
     {
      "bearerAuth": []
     }
    ],
    "responses": {
     "200": {
      "description": "Success",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "properties": {
          "gateway_models": {
           "type": "array",
           "items": {
            "type": "object",
            "properties": {
             "id": {
              "type": "string"
             },
             "brand": {
              "type": "string"
             },
             "model": {
              "type": "string"
             },
             "label": {
              "type": "string"
             },
             "transport": {
              "type": "string",
              "enum": [
               "mqtt",
               "http"
              ],
              "description": "mqtt: per-gateway MQTT credentials; http: POST /ingest/gateways/{id}/packets with HTTP Basic. minew-mg4 is listed as mqtt and can also upload over HTTP."
             },
             "description": {
              "type": "string"
             },
             "logo": {
              "type": "string"
             },
             "verified": {
              "type": "boolean"
             },
             "image": {
              "type": "string",
              "description": "Product photo path (manufacturer image, identification only)"
             }
            }
           }
          },
          "device_profiles": {
           "type": "array",
           "items": {
            "type": "object",
            "properties": {
             "id": {
              "type": "string"
             },
             "brand": {
              "type": "string"
             },
             "model": {
              "type": "string"
             },
             "label": {
              "type": "string"
             },
             "radio": {
              "type": "string",
              "enum": [
               "ble",
               "any"
              ]
             },
             "description": {
              "type": "string"
             },
             "image": {
              "type": "string"
             },
             "kinds": {
              "type": "array",
              "items": {
               "type": "string"
              }
             },
             "metrics": {
              "type": "array",
              "items": {
               "type": "string"
              }
             },
             "verified": {
              "type": "boolean"
             },
             "info_name": {
              "type": "string"
             },
             "notes": {
              "type": "string"
             },
             "wearable": {
              "type": "boolean"
             },
             "button": {
              "type": "boolean"
             },
             "occupancy": {
              "type": "boolean",
              "description": "PIR sensor: the motion metric drives occupied / vacant events (vacant after 300 s without motion, raised by the periodic worker)."
             },
             "door": {
              "type": "boolean",
              "description": "Door contact: the door metric (1 open, 0 closed) drives door_open / door_closed. For minew-s4 it comes from a taught signal; no frame decoder exists."
             }
            }
           }
          },
          "verification": {
           "type": "string"
          },
          "alerts_shadow": {
           "type": "boolean"
          }
         }
        }
       }
      }
     },
     "default": {
      "description": "Rejected; 400 invalid input, 401 credentials, 403 authorization/origin or detail=password_change_required, 404 inaccessible resource, 409 conflict/quota, 429 rate limit, 500 internal.",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/catalog/zigbee": {
   "get": {
    "summary": "Search the Zigbee2MQTT device catalog (zigbee-herdsman-converters, MIT): every supported model with the category Aether shows it as",
    "operationId": "get__api_v1_catalog_zigbee",
    "security": [
     {
      "bearerAuth": []
     }
    ],
    "parameters": [
     {
      "name": "q",
      "in": "query",
      "schema": {
       "type": "string",
       "maxLength": 100
      },
      "description": "Every word must match vendor, model, description, Zigbee model id or white label"
     },
     {
      "name": "vendor",
      "in": "query",
      "schema": {
       "type": "string",
       "maxLength": 64
      }
     },
     {
      "name": "category",
      "in": "query",
      "schema": {
       "type": "string",
       "maxLength": 32
      }
     },
     {
      "name": "limit",
      "in": "query",
      "schema": {
       "type": "integer",
       "minimum": 1,
       "maximum": 200,
       "default": 50
      }
     },
     {
      "name": "vendors",
      "in": "query",
      "schema": {
       "type": "string",
       "enum": [
        "1"
       ]
      },
      "description": "Also return the list of vendors"
     }
    ],
    "responses": {
     "200": {
      "description": "Success",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "properties": {
          "items": {
           "type": "array",
           "items": {
            "type": "object",
            "properties": {
             "vendor": {
              "type": "string"
             },
             "model": {
              "type": "string"
             },
             "description": {
              "type": "string"
             },
             "zigbee_model": {
              "type": "array",
              "items": {
               "type": "string"
              }
             },
             "white_label": {
              "type": "array",
              "items": {
               "type": "string"
              }
             },
             "category": {
              "type": "string"
             },
             "features": {
              "type": "array",
              "items": {
               "type": "string"
              }
             },
             "sos": {
              "type": "boolean"
             },
             "dynamic": {
              "type": "boolean"
             }
            }
           }
          },
          "total": {
           "type": "integer"
          },
          "source": {
           "type": "string"
          },
          "version": {
           "type": "string"
          },
          "license": {
           "type": "string"
          },
          "homepage": {
           "type": "string"
          },
          "notice": {
           "type": "string"
          },
          "vendors": {
           "type": "array",
           "items": {
            "type": "string"
           }
          }
         }
        }
       }
      }
     },
     "400": {
      "description": "Query too long or limit out of range"
     },
     "401": {
      "description": "Not authenticated"
     }
    }
   }
  },
  "/api/v1/gateways": {
   "get": {
    "summary": "List tenant gateways",
    "operationId": "get__api_v1_gateways",
    "security": [
     {
      "bearerAuth": []
     }
    ],
    "responses": {
     "200": {
      "description": "Success",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "properties": {
          "items": {
           "type": "array",
           "items": {
            "$ref": "#/components/schemas/Gateway"
           }
          }
         },
         "required": [
          "items"
         ],
         "additionalProperties": false
        }
       }
      }
     },
     "default": {
      "description": "Rejected; 400 invalid input, 401 credentials, 403 authorization/origin or detail=password_change_required, 404 inaccessible resource, 409 conflict/quota, 429 rate limit, 500 internal.",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    }
   },
   "post": {
    "summary": "Create gateway; owner/admin only",
    "operationId": "post__api_v1_gateways",
    "security": [
     {
      "bearerAuth": []
     }
    ],
    "responses": {
     "201": {
      "description": "Success",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/GatewayCreated"
        }
       }
      }
     },
     "default": {
      "description": "Rejected; 400 invalid input, 401 credentials, 403 authorization/origin or detail=password_change_required, 404 inaccessible resource, 409 conflict/quota, 429 rate limit, 500 internal.",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/NewGateway"
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "securitySchemes": {
   "bearerAuth": {
    "type": "http",
    "scheme": "bearer",
    "bearerFormat": "JWT"
   },
   "gatewayBasic": {
    "type": "http",
    "scheme": "basic",
    "description": "Gateway UUID as username and one-time-displayed token as password; HTTPS required outside loopback."
   },
   "refreshCookie": {
    "type": "apiKey",
    "in": "cookie",
    "name": "__Host-aether_refresh"
   },
   "developmentRefreshCookie": {
    "type": "apiKey",
    "in": "cookie",
    "name": "aether_refresh_dev"
   }
  },
  "schemas": {
   "Error": {
    "type": "object",
    "properties": {
     "error": {
      "type": "string"
     },
     "request_id": {
      "type": "string"
     },
     "detail": {
      "type": "string",
      "enum": [
       "password_change_required"
      ],
      "description": "Machine-readable reason, currently only password_change_required: the session was issued an initial or reset password and must replace it through POST /api/v1/auth/password before any other route answers. Sent with 403."
     }
    },
    "required": [
     "error",
     "request_id"
    ],
    "additionalProperties": false
   },
...
   "Account": {
    "type": "object",
    "properties": {
     "tenant_id": {
      "type": "string",
      "format": "uuid"
     },
     "user": {
      "type": "object",
      "properties": {
       "id": {
        "type": "string",
        "format": "uuid"
       },
       "email": {
        "type": "string"
       },
       "name": {
        "type": "string"
       }
      },
      "required": [
       "id",
       "email",
       "name"
      ],
      "additionalProperties": false
     }
    },
    "required": [
     "tenant_id",
     "user"
    ],
    "additionalProperties": false
   },
...
  }
 }
}